+++ src/lib/RedirectService.js
@@ -1,28 +1,7 @@
-export default class RedirectService {
-  constructor() {
-    this.routes = {
-      CLIENT_APP: {
-        base: 'https://volunteero-front-end-base.herokuapp.com/',
-      },
-    };
-  }
+/**
+ * This is a translation module to use 'export'  and 'import' module syntax
+ */
 
-  /**
-   * Returns a route with the query string from internal route table
-   * returns a valid url if found and null if the route with key is not known
-   * @param {String} key
-   * @return {String | null}
-   */
-  getUrl(key, ...queryParams) {
-    const route = this.routes[key];
-    console.log(route);
-    // TODO: use a better existance check
-    if (route) {
-      const queryArray = queryParams.map(q => `${q.key}=${q.value}`);
-      const queryString = (queryArray.length > 0) ? queryArray.join('&') : '';
-      const querySuffix = (queryString === '') ? '' : `?${queryString}`;
-      return `${route.base}${querySuffix}`;
-    }
-    return null;
-  }
-}
+const RedirectService = require('../../shared/RedirectService');
+
+export default RedirectService;
